# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LOGIK GAME API",
        "description": "API backend du jeu télévisé LOGIK GAME — gestion des sessions, manches, questions, joueurs, temps réel et projection.",
        "contact": {
            "name": "NerdX",
            "email": "contact@nerdx.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/admin/sessions/{session}/preselection-questions": {
            "get": {
                "tags": [
                    "Preselection Questions"
                ],
                "summary": "Lister les questions de pré-sélection",
                "operationId": "c63fa8a9bed0bf604e7c49d51bc9b1f2",
                "parameters": [
                    {
                        "name": "session",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des questions"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Preselection Questions"
                ],
                "summary": "Créer une question de pré-sélection",
                "operationId": "920c5dba9017dc257c34bebb27c45935",
                "parameters": [
                    {
                        "name": "session",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "text",
                                    "answer_type"
                                ],
                                "properties": {
                                    "text": {
                                        "type": "string",
                                        "maxLength": 2000
                                    },
                                    "answer_type": {
                                        "type": "string",
                                        "enum": [
                                            "qcm",
                                            "text",
                                            "number"
                                        ]
                                    },
                                    "correct_answer": {
                                        "description": "Obligatoire pour text/number. Pour qcm, auto-dérivé du choix is_correct=1.",
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "duration": {
                                        "description": "Durée en secondes (défaut: 30)",
                                        "type": "integer",
                                        "example": 30
                                    },
                                    "display_order": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "number_is_decimal": {
                                        "description": "Uniquement pour answer_type=number",
                                        "type": "boolean",
                                        "nullable": true
                                    },
                                    "media_file": {
                                        "description": "Fichier média (image/vidéo/audio, max 50Mo)",
                                        "type": "string",
                                        "format": "binary",
                                        "nullable": true
                                    },
                                    "media_type": {
                                        "description": "Auto-détecté si non fourni",
                                        "type": "string",
                                        "enum": [
                                            "none",
                                            "image",
                                            "video",
                                            "audio"
                                        ]
                                    },
                                    "choices": {
                                        "description": "Obligatoire si answer_type=qcm (4 à 6 choix)",
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "label": {
                                                    "type": "string"
                                                },
                                                "is_correct": {
                                                    "type": "boolean"
                                                },
                                                "display_order": {
                                                    "type": "integer",
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Question créée"
                    },
                    "422": {
                        "description": "Validation échouée"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/sessions/{session}/preselection-questions/{preselectionQuestion}": {
            "get": {
                "tags": [
                    "Preselection Questions"
                ],
                "summary": "Détail question pré-sélection",
                "operationId": "c5016ce1a5afec2a3742ce0ae762048c",
                "parameters": [
                    {
                        "name": "session",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "preselectionQuestion",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détail question"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Preselection Questions"
                ],
                "summary": "Modifier question pré-sélection",
                "description": "Envoyer via POST avec le champ `_method=PUT` pour supporter les fichiers (multipart/form-data).",
                "operationId": "c264fce4ded3240a8aabcfa2a3e33158",
                "parameters": [
                    {
                        "name": "session",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "preselectionQuestion",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "_method": {
                                        "description": "Workaround navigateur pour PUT multipart",
                                        "type": "string",
                                        "example": "PUT"
                                    },
                                    "text": {
                                        "type": "string",
                                        "nullable": true,
                                        "maxLength": 2000
                                    },
                                    "answer_type": {
                                        "type": "string",
                                        "nullable": true,
                                        "enum": [
                                            "qcm",
                                            "text",
                                            "number"
                                        ]
                                    },
                                    "correct_answer": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "duration": {
                                        "description": "Durée en secondes",
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "display_order": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "number_is_decimal": {
                                        "type": "boolean",
                                        "nullable": true
                                    },
                                    "media_file": {
                                        "description": "Remplace le fichier existant (max 50Mo)",
                                        "type": "string",
                                        "format": "binary",
                                        "nullable": true
                                    },
                                    "media_type": {
                                        "type": "string",
                                        "nullable": true,
                                        "enum": [
                                            "none",
                                            "image",
                                            "video",
                                            "audio"
                                        ]
                                    },
                                    "remove_media": {
                                        "description": "Supprimer le média existant sans en uploader un nouveau",
                                        "type": "boolean"
                                    },
                                    "choices": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "label": {
                                                    "type": "string"
                                                },
                                                "is_correct": {
                                                    "type": "boolean"
                                                },
                                                "display_order": {
                                                    "type": "integer",
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Question modifiée"
                    },
                    "422": {
                        "description": "Validation échouée"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Preselection Questions"
                ],
                "summary": "Supprimer question pré-sélection",
                "operationId": "547ef0b94eb76454a3689897fa0932f7",
                "parameters": [
                    {
                        "name": "session",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "preselectionQuestion",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Question supprimée"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Token Sanctum admin (obtenu via POST /admin/login)",
                "bearerFormat": "Sanctum Token",
                "scheme": "bearer"
            },
            "playerToken": {
                "type": "apiKey",
                "description": "Token d'accès joueur (64 caractères, reçu après sélection)",
                "name": "X-Player-Token",
                "in": "header"
            }
        }
    },
...
}